Added autogenerated (inter)key routing and non-hotswapability
With these alterations ergogen will generate a non-hotswap key on top of
the hot swap one for better flexibility. This patch also adds basic routing
within the key itself, and between keys with a known location. This reduces
the amount of wiring that needs to be done and keeps the look of each key
the same.